Alpha Preparation

+++ Nominax/Include/Nominax/Foundation/Record.hpp
@@ -206,9 +206,8 @@
 #pragma once
 
 #include <array>
-#include <type_traits>
-
 #include <cstdint>
+#include <type_traits>
 
 namespace Nominax::Foundation
 {
@@ -491,7 +490,7 @@ namespace Nominax::Foundation
 
 	constexpr auto Record::Padding() -> Record
 	{
-		return Record { 0xFF'FF'FF'FF'FF'FF'FF'FF };
+		return Record { static_cast<std::uint64_t>(0xFF'FF'FF'FF'FF'FF'FF'FF) };
 	}
 
 	static_assert(sizeof(float) == sizeof(std::int32_t));
